refactor: Disable crate for non-wasm in a different way
